src/main/java/de/variantsync/studies/sync/util/ShellExecutor.java
package de.variantsync.studies.sync.util;

import de.variantsync.evolution.util.Logger;
import de.variantsync.evolution.util.functional.Result;
import de.variantsync.evolution.util.functional.Unit;
import de.variantsync.studies.sync.exception.ShellException;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.nio.file.Path;
import java.util.function.Consumer;

public class ShellExecutor {

    public ShellExecutor(Consumer<String> outputReader) {

    }

    public ShellExecutor(Path resourcesDir, Consumer<String> outputReader) {

    }

    public Result<Unit, ShellException> execute(IShellCommand command) {
        return Result.Success(Unit.Instance());
    }


    private static class StreamGobbler implements Runnable {
        private InputStream inputStream;
        private Consumer<String> consumer;

        public StreamGobbler(InputStream inputStream, Consumer<String> consumer) {
            this.inputStream = inputStream;
            this.consumer = consumer;
        }

        @Override
        public void run() {
            try(BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream))) {
                reader.lines().forEach(consumer);
            } catch (IOException e) {
                Logger.error("Exception thrown while reading stream of Shell command.", e);
            }
        }
    }
}